+++ R/brm_simulate_simple.R
@@ -63,10 +63,19 @@ brm_simulate_simple <- function(
   assert_pos(hyper_tau, message = "hyper_tau must be 1 positive number")
   assert_pos(hyper_lambda, message = "hyper_lambda must be 1 positive number")
   patients <- tibble::tibble(
-    group = paste0("group_", rep(seq_len(n_group), each = n_patient)),
-    patient = paste0("patient_", seq_len(n_group * n_patient))
+    group = paste0(
+      "group_",
+      zero_pad_integers(rep(seq_len(n_group), each = n_patient))
+    ),
+    patient = paste0(
+      "patient_",
+      zero_pad_integers(seq_len(n_group * n_patient))
+    )
   )
-  levels_time <- paste0("time_", seq_len(n_time))
+  levels_time <- paste0(
+    "time_",
+    zero_pad_integers(seq_len(n_time))
+  )
   grid <- tidyr::expand_grid(patients, time = levels_time)
   model_matrix <- model.matrix(~ 0 + group + time, data = grid)
   beta <- stats::rnorm(n = ncol(model_matrix), mean = 0, sd = hyper_beta)
@@ -93,8 +102,8 @@ brm_simulate_simple <- function(
     patient = "patient",
     covariates = character(0L),
     missing = NULL,
-    reference_group = "group_1",
-    reference_time = "time_1"
+    reference_group = min(data$group),
+    reference_time = min(data$time)
   )
   parameters <- list(
     beta = beta,