+++ R/utils_data.R
@@ -4,3 +4,7 @@ unname_df <- function(x) {
   }
   x
 }
+
+zero_pad_integers <- function(x) {
+  sprintf(paste0("%0", max(nchar(as.character(x))), "d"), as.integer(x))
+}