+++ NEWS.md
@@ -4,6 +4,7 @@
 * Add new functions for nuanced simulations: `brm_simulate_outline()`, `brm_simulate_continuous()`, `brm_simulate_categorical()` (#3).
 * In `brm_model()`, remove rows with missing responses. These rows are automatically removed by `brms` anyway, and by handling by handling this in `brms.mmrm`, we avoid a warning.
 * Add subgroup analysis functionality (#18).
+* Zero-pad numeric indexes in simulated data so the levels sort as expected.
 * In `brm_data()`, deprecate `level_control` in favor of `reference_group`.
 * In `brm_data()`, deprecate `level_baseline` in favor of `reference_time`.